requirements.txt
streamlit
faiss-cpu
numpy
requests
beautifulsoup4
python-dotenv
openai